openacc: prevent pointer-to-int-cast warning
Pass SCOREP_RegionHandle via uintptr_t cast to and from OpenACC
runtime's void* tool_info member.

(cherry picked from commit 6034c4e1024a0682e6d2fbc96b6c4e963c7de386)

+++ src/adapters/openacc/scorep_openacc_event.c
@@ -123,8 +123,7 @@ scorep_openacc_handle_enter_region( acc_prof_info*  profInfo,
                                           profInfo->src_file );
 
     // pass Score-P region handle to corresponding leave
-    // causes compiler warning!!!
-    eventInfo->other_event.tool_info = ( void* )( long )region_handle;
+    eventInfo->other_event.tool_info = ( void* )( uintptr_t )region_handle;
 
     // add an attribute, if this is an implicit region
     if ( eventInfo->other_event.implicit )
@@ -165,7 +164,7 @@ handle_leave_region( acc_prof_info*  profInfo,
 
     if ( eventInfo )
     {
-        regionHandle = ( uint32_t )eventInfo->other_event.tool_info;
+        regionHandle = ( uintptr_t )eventInfo->other_event.tool_info;
     }
 
     if ( regionHandle )
